pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "es-hotzone-trader"
version = "0.1.0"
description = "ES Hot-Zone Day Trading CLI System"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Trader", email = "trader@example.com"}
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "websockets>=12.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pytz>=2023.3",
    "aiohttp>=3.9.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
gui = [
    "PySide6>=6.6.0",
    "pyqtgraph>=0.13.3",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
es-trade = "src.cli.commands:main"
es-trade-gui = "src.gui.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
pythonpath = ["."]
